+++ src/components/MainPage/Header/Administration/AdminandDirectorates.css
@@ -35,8 +35,7 @@
   font-weight: bold;
   text-align: center;
   color: #04AA6D;
-}
-
+} 
 .ADimg {
   height: 300px;
   width: 300px;
