+++ src/components/MainPage/Header/Administration/Registrar/Registrar.js
@@ -1,137 +1,33 @@
 import React from "react";
-import "../Administration/AdminandDirectorates.css"; // Import your CSS file here
-import { Link, useLocation } from "react-router-dom";
-function FacultyDevelopmentCell() {
-  const location = useLocation();
+import registrar from "../../../../../assets/AdminDirectoratesImages/registrar.jpeg";
+import "../AdminandDirectorates.css"; // Import your CSS file here
+
+function Registrar() {
+ 
   return (
     <div className="mainContainer">
-      <div className="allLeftMenu">
-        {/* Navigation links for Chancellor, Vice Chancellor, and Registrar */}
-        <Link
-          to="/AcademicAudit"
-          className={`menuButton ${
-            location.pathname === "/AcademicAudit" ? "active" : ""
-          }`}
-        >
-          Director of Academic Audit
-        </Link>
-        <Link
-          to="/AcademicPlanning"
-          className={`menuButton ${
-            location.pathname === "/AcademicPlanning" ? "active" : ""
-          }`}
-        >
-          Director of Academic Planning
-        </Link>
-        <Link
-          to="/Admissions"
-          className={`menuButton ${
-            location.pathname === "/Admissions" ? "active" : ""
-          }`}
-        >
-          Director of Admissions
-        </Link>
-        <Link
-          to="/AlumniRelations"
-          className={`menuButton ${
-            location.pathname === "/AlumniRelations" ? "active" : ""
-          }`}
-        >
-          Director of Alumni Relations
-        </Link>
-        <Link
-          to="/Empowerment"
-          className={`menuButton ${
-            location.pathname === "/Empowerment" ? "active" : ""
-          }`}
-        >
-          Empowerment of Women & Grievances
-        </Link>
-        <Link
-          to="/Evaluation"
-          className={`menuButton ${
-            location.pathname === "/Evaluation" ? "active" : ""
-          }`}
-        >
-          Evaluation
-        </Link>
-        <Link
-          to="/FacultyDevelopmentCell"
-          className={`menuButton ${
-            location.pathname === "/FacultyDevelopmentCell" ? "active" : ""
-          }`}
-        >
-          Faculty Development Cell
-        </Link>
-        <Link
-          to="/ForeignUniversityRelations"
-          className={`menuButton ${
-            location.pathname === "/ForeignUniversityRelations" ? "active" : ""
-          }`}
-        >
-          Foreign University Relations
-        </Link>
-        <Link
-          to="/IQAC"
-          className={`menuButton ${
-            location.pathname === "/IQAC" ? "active" : ""
-          }`}
-        >
-          Director of IQAC
-        </Link>
-        <Link
-          to="/Placements"
-          className={`menuButton ${
-            location.pathname === "/Placements" ? "active" : ""
-          }`}
-        >
-          Director of IIIT & Placements
-        </Link>
-        <Link
-          to="/Research"
-          className={`menuButton ${
-            location.pathname === "/Research" ? "active" : ""
-          }`}
-        >
-          Director of Research & Development
-        </Link>
-        <Link
-          to="/SportsAdministration"
-          className={`menuButton ${
-            location.pathname === "/SportsAdministration" ? "active" : ""
-          }`}
-        >
-          Director of University Games & Sports Administration
-        </Link>
-      </div>
       <div className="allRightContent">
-        <div className="allRightContentHeading">
-          Director of Faculty Development Cell
-        </div>
+        <div className="allRightContentHeading">Registrar</div>
         <br></br>
         <div className="allRightContentProfile">
           {/* Profile information */}
           <center>
             <div className="profileImage">
-              <img
-                src="https://jntugv.edu.in/images/9266a0f9fc226ffbf57a3beb4c6b6388/profilephotos/registrar.jpeg"
-                alt="Profile"
-              />
+              <img src={registrar} alt="Profile" />
             </div>
             <div className="profileName">
-              <h3>Dr.G.Jaya Suma</h3>
-              <h3>Director of Faculty Development Cell</h3>
+              <h3>Prof.Dr.G.Jaya Suma</h3>
+              <h3>Registrar</h3>
             </div>
             <div className="profileDesignation">
               <h4 className="regularStyleH4">
-                Professor of Dept. of Information Technology
+                Professor of Dept.of information Technology
               </h4>
             </div>
           </center>
           <hr />
           {/* Profile description */}
-          <p>Email: dfdp@jntugv.edu.in</p>
-          <br></br>
+          <p>Email: registrar@jntugv.edu.in</p>
 
           <p>
             Dr. G. Jaya Suma, present working as a Professor in Department of
@@ -215,4 +111,4 @@ function FacultyDevelopmentCell() {
   );
 }
 
-export default FacultyDevelopmentCell;
+export default Registrar;
