+++ src/components/MainPage/Header/Administration/ViceChancellor/ViceChancellor.js
@@ -1,40 +1,9 @@
 import React from "react";
-import vc from "../../../../assets/AdminDirectoratesImages/vc.png";
-import "./AdminandDirectorates.css"; // Import your CSS file here
-
-import { Link, useLocation } from "react-router-dom";
-
+import vc from "../../../../../assets/AdminDirectoratesImages/vc.png";
+import "../AdminandDirectorates.css" // Import your CSS file here
 function ViceChancellor() {
-  const location = useLocation();
   return (
     <div className="mainContainer">
-      <div className="allLeftMenu">
-        {/* Navigation links for Chancellor, Vice Chancellor, and Registrar */}
-        <Link
-          to="/Chancellor"
-          className={`menuButton ${
-            location.pathname === "/Chancellor" ? "active" : ""
-          }`}
-        >
-          Chancellor
-        </Link>
-        <Link
-          to="/ViceChancellor"
-          className={`menuButton ${
-            location.pathname === "/ViceChancellor" ? "active" : ""
-          }`}
-        >
-          Vice Chancellor
-        </Link>
-        <Link
-          to="/Registrar"
-          className={`menuButton ${
-            location.pathname === "/Registrar" ? "active" : ""
-          }`}
-        >
-          Registrar
-        </Link>
-      </div>
       <div className="allRightContent">
         <div className="allRightContentHeading">Vice-Chancellor</div>
         <br></br>
